Adjust control code.
We were getting issues when using the built in map function. This may
be due to the implicit conversion that was occurring from type double to
type long. To address this, we created a new function that is identical
to the Arduino map function but with double data types instead of longs.